server/requirements.txt
altair==5.5.0
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
beautifulsoup4==4.14.2
billiard==4.2.2
blinker==1.9.0
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
curl_cffi==0.13.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fredapi==0.5.2
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.45
gunicorn==25.1.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
MarkupSafe==3.0.3
multitasking==0.0.12
narwhals==2.10.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
peewee==3.18.2
pillow==12.0.0
platformdirs==4.5.0
plotly==6.4.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
scipy==1.16.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
sqlparse==0.5.3
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
yfinance==0.2.66